requirements.txt
absl-py==0.7.1
astor==0.7.1
Django==2.1.7
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
image==1.5.27
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.0.1
mock==2.0.0
numpy==1.16.2
pbr==5.1.3
Pillow==5.4.1
pkg-resources==0.0.0
protobuf==3.7.0
PyPrind==2.11.2
pytz==2018.9
scipy==1.2.1
six==1.12.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
Werkzeug==0.15.1
